scripts/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
Pillow>=10.0.0
PyNaCl>=1.5.0
